withErrorBoundary (#8629)
Summary:
If cord crashes it should not crash everything.
Every withCord now adds an Error Boundary.

The error boundary catch any error and will render a replaceable ErrorFallback (that renders `null`).
 
The fact that all our components do that may not be the clever thing, but it is simple.


Test Plan:
In testbed get the composer to throw on render.
See the error.
Replace the ErrorFallback and see it as well.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/5c4aea5f-3734-4671-88be-159252626026.png)


Reviewers:

monorepo-commit: d5d9c586fcb280cbb3e0be923b9ef47c344b2195