MediaModal: Drill user/date to banner
The MediaModal shows a banner a the top with some information.
However, this information was never passed down from `Message`.
This resulted in an empty banner (because somewhere we render `banner &&`).

Also fixed the TODO so that the user/date information is shown in the UI (instead of `{{ user.name }}` `{{ date }}`).

Test Plan:
Open a MediaModal, check it now has a banner with buttons, and the buttons work.


Reviewers: Netceer, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8674

monorepo-commit: e82629c34bfbeab3838b92384c811520bf73bbc7