Fix the webhook helper functions (#8662)
The server library webhook helper functions didn't actually work
properly/usefully as written, so make them work properly.

The validation function had its input types set up such that it
couldn't actually verify the signature: it expected a structured
object (eg, after JSON parsing), but the signature is computed over
the raw bytes of the request.  It also took a headers function, but
that doesn't really work in all cases, so just take the two headers we
need as strings.

The body parsing function tried to return the event payload of the
event, but (A) it didn't actually do that, it returned the wrapper, so
none of the fields worked, and (B) returning the message payload isn't
very helpful because you don't know what kind of message it is.

This is backwards-incompatible, but since the previous functions
didn't work properly, I don't think anyone is using them, so I changed
functions and removed types willy-nilly here.

Test Plan:

Copied-and-pasted the source of these functions into the AI Quiz
webhook handler and replaced all the current code with:

```
const rawBody = await req.text();

validateWebhookSignature(
  rawBody,
  req.headers.get('X-Cord-Timestamp'),
  req.headers.get('X-Cord-Signature'),
  CORD_API_SECRET,
);

const data = parseWebhookBody<'thread-message-added'>(rawBody);
```

This is what we want the code to calling code to look like, and it all
works and validates signatures successfully.

monorepo-commit: d848651379129b7f174c3a47bc915eed526269cb