Adding PresenceObserver docs  (#8660)
## Summary

Adding docs for PresenceObserver. Most things are the same as before, but we need to exclude vanilla js and also move the description of the properties to the types file, instead of having it in the docs page. I could have separated this into PresenceObserver3 and PresenceObserver4 as I did with message, and I can do it if you want me too, but I think there is more overlap in all the text than the other two.

## Test plan

Uploading Screen Recording 2024-04-05 at 16.24.23.mov…

monorepo-commit: 0d6465e75299f6df6f7c79d0a56b9169a7510e5f