OptionsMenu items are single items instead of grouped. (#8687)
Summary:
It was bad that Thread/Message actions were one single item rendering
multiple items. it was hardly replacable.
But we cannot have a React component returning an array, so it becomes a hook.
|before|after|
|----|----|
|![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/b8e7c153-4879-430b-856a-899b7fb79a9f.png)|![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/2809ec8c-fa53-46a5-a8e5-571b1553502f.png)|


This extracts creating the items into a hook, so we can use it in the actual OptionsMenu
Test Plan:
Test bed thread has options menu and resolve menu item.
And Menu's items are properly flatten.

Reviewers:

monorepo-commit: feb54427ed284e61c04d6459853ee607529aaac1