Rename presence.observeLocationData to observePresence (#8693)
We've had a lot of success with this style of naming in the thread
context (`observeThreads`, etc.), so rename
`presence.observeLocationData` to `observePresence` to match.  The old
name continues to work.

The docs don't even need to change URLs, because their URL is
`/getPresent`

Test Plan:

Called both APIs from a browser and then changed presence, both
updated effectively simultaneously.

monorepo-commit: 86bc025e87ced7ee87b3197120ab8e8b5aba0eb3