Upgrade phosphor icons (#8720)
Upgrade from `phosphor-react` to `@phosphor-icons/react`.  This should
deal with an issue a client was hitting where `phosphor-react` defined
a symbol named `Infinity`, which is also the name of the constant for
mathematical infinity and some build systems objected.

Because the new version of `@phosphor-icons/react` imports files with
no file extension, it isn't actually compatible with our tsconfig
settings (which are based on the absolutely most restrictive version
possible).  To solve that, send everything through a file with
`@ts-nocheck` in it, because the actual code operates properly at
runtime.

Test Plan:

Build the opensource packages, they all built properly with no errors.

Loaded the docs, all the icons show up in the proper places.

monorepo-commit: e068f83e60085c667fbc0ebf7ec12f1d780cb1e6