Deprecate three non-useful thread read options (#8732)
We have three options on `ThreadObserverOptions` that were there to
support the fact that reading a thread that didn't exist used to
create it as a side effect, and so we needed a place to supply the
required properties of a thread.  We don't do that anymore, so we no
longer need them.

Two of them (`threadName` and `location`) are just functionless now,
so deprecate them with no replacement.  The third actually also
functions as a filter (if you're trying to read a thread not in that
group, you get no thread back), so redirect uses of those to the
equivalent filter argument.


Test Plan:

Passing a groupID to `observeThread` still effectively filters by that
group.

monorepo-commit: 592a6ffd5eddb50330da34bfe7c03353b229e192