Release SDK 1.39.0
AFAIK contains just a fix for `validateWebhookSignature`, but I want
that fix. It's also technically a BC break, but the function is
undocumented (though we should fix that), so not worried about it.


Test Plan: Scripts are all happy, CI is happy, changelog looks good.

Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/8747

monorepo-commit: c54c5dfed83dcc318ccd87a63ca76c32513e1933