Versioned CSS: always pull from app.cord.com
The `upload-CSS` script always and only uploads to `app.cord.com`'s bucket -- and it makes sense to stay that way: we bundle and upload the CSS as part of publishing a new npm package. We don't need to have a "staging" version of the CSS.

Notice this because Clack uses the npm package, but uses staging. It linked to a CSS file that didn't exist in staging (but did in prod).

Test Plan:
Refresh Clack, can now see the CSS file served correctly.


Reviewers: lazybean, elenaperezrioja, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8754

monorepo-commit: 3f121490e2efe3eb908754ca2ea29538e19b0688