Multiple changes to props (#8757)
## Summary
I've gone through all our components checking that everything has a classname and that we're passing classnames down.
Some of the components didn't get StyleProps, so switching to that. I also prefer the ...restProps instead of passing className and style in case we extend what is passed in StyleProps in the future.

monorepo-commit: 2075fabb0045d7202245cc0dd5615b9473c0372f