ScrollContainer: Better onScrollToEdge logic
`element.scrollTop`, or `element.scrollTo` trigger a `scroll` event just like an actual user scrolling.

If we implemented infinite scrolling, we would get to the bottom, fetch new items and autoscroll to the newest.. Which will trigger a `scroll` event and keep scrolling down till the end of time.

This PR makes `onScrollToEdge` be triggered only if it's the first time we bumped into an edge, effectively preventing the infinite-fetch-scroll loop.

Test Plan:
Add a console.log to `onScrollToEdge`.
Check it gets triggered only if the edge changes.

Implement a `fetchMore` on scroll to edge. Check we fetch more only once. But if we scroll to the bottom again, we `fetchMore`.


Reviewers: Netceer, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8790

monorepo-commit: a9b57ab1ca4267ef6545321b46107c0bd5299c98