Add missing message grid area (#8786)
Summary:
Message grid is complicated enough without being broken.
If you were changing the position of avatar or timestamp in the grid
template it would not do what you expected because the actual avatar and
timestamp are not assigned to that grid aera. :face-palm:

Test Plan:
In doc, change the grid template. Magic.

Reviewers:

monorepo-commit: 946600c6fbeeb1e420a7b4fec4d67c164eccc39e