ThreadLayout: threadSeenBy inside ScrollContainer
That's where it should be. That's where it is in ui3 world, too.

Test Plan:
Send lots of messages in a thread, open a second tab with a different user, switch back to the first tab, check I see the ThreadSeenBy component inside the scroll container.


Reviewers: Netceer, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8810

monorepo-commit: e67475a86a3fa87e0ea3794640bd746deba08a33