ScrollContainer: Add onOverflowChange
Knowing when the overflow changes can be useful to e.g. "keep fetching messages until we fill the screen -- i.e. until we overflow"

Test Plan:
Add message until we overflow, check we get the callback `true`.
Delete a bunch of messages, check we get the callback `false`.

Set scroll container's maxHeight to 20 (so that as soon as we render it, we have an overflow).
I get the onOverflowChange -> true.


Reviewers: Netceer, lazybean

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/8807

monorepo-commit: 67344f6a76658a7676afd370850bea1a36ccda3e