Introduce a "flag" attribute type, use it for autofocus and disabled
Autofocus (and I think disabled) mean something special in the HTML
spec. And per spec, it only matters if the attribute is present, not
what it's set to (though it's always supposed to be set to `""`). In
particular, `autofocus=false` isn't a thing -- that's the same as
`autofocus`.

For these, introduce a special type of boolean called a "flag", which is
omitted instead of set to false, and move autofocus and disabled over to
it. (I don't move other things since it prevents defaulting to true, and
for most of these, they are our own things that we are allowed to
interpret however we like -- only these are special in HTML.)


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/8802

monorepo-commit: d2980870a17d423ba8078928451b0e791cf75d7b