Message components: pass refs and classnames
We realised that only 2/4 elements that the message renders actually passed down refs and classnames. Adding the remaining 2 plus the SendComposer.

This is useful because e.g. if I want to style the edit composer, I can use a classname that I pass to it.

Test Plan:
Load Thread, Composer and Message in the new Docs. No errors in the console, I can now pass a className to `Message` and it gets passed down correctly to all the components below (e.g. the Tombstone)


Reviewers: lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8830

monorepo-commit: 307776187c656385afe04a1111be20b422af394e