Message 4.0: Make it easier to style
By default, the message will keep growing as much as it wants.
This change  makes is so that if you set a `max-height` on the Message, and you have a very long message, a scrollbar will be added to the message text alone, which we thought was the most reasonable UX.
Before this change, setting a max-height would overflow the composer.

Test Plan:
Load Message in the docs, type a long message, check it looks OK.
Now add a max-height to it, check it looks good.


| BEFORE | AFTER |
|-|-|
| <img width="643" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/793a5260-a357-4091-86a5-9388eb47c649"> | <img width="710" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/a032083e-2b6e-4755-a3de-17b7ab5d4c6d">|
| <img width="643" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/d4b6d9f5-b175-45ec-95f2-466a79f4bb18"> | <img width="701" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/f5179220-958b-469c-bf0f-3c205d5ba630"> |






Reviewers: lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8833

monorepo-commit: 5ff2da9bef65abfd41d884f22dee4e1086dcb311