Use MessageContent type in message definition (#8866)
We have an exported type for MessageContent now, so replace the
existing type of `object[]` with the new type.  This also lets us
remove a bunch of typecasts and other handling oddities because we
required proper typing in components and things.

Test Plan:

Everything compiles, docs look the same as they did before

monorepo-commit: ea716e3f3bdacc3a5c76099f537b59c4e4c0114c