Support supplying user data via context (#8789)
Add a new context, UserDataContext, that can be used to supply user
data rather than having components call the JS API to get it.  Also
add useComponentUserData, which is a replacement for useUserData that
consults this context and uses it preferentially.  This is mostly
useful for SSR, but also is useful for some advanced customization use
cases.  Finally, clean up a couple props.

Test Plan:

Added to the docs:

```
<UserDataContext.Provider
  value={{
    users: {
      'f2e73e38-89fe-49ec-b5c1-95477c3ad5b4': {
        id: 'f2e73e38-89fe-49ec-b5c1-95477c3ad5b4',
        name: 'Not Dave',
        shortName: null,
        displayName: 'Not Dave',
        secondaryDisplayName: 'Not Dave',
        metadata: {},
        profilePictureURL: null,
      },
    },
  }}
>
  <experimental.Message message={threadData.firstMessage!} />
</UserDataContext.Provider>
```

The user data for that user was overwritten to show "Not Dave" in the
author, avatar, etc.

monorepo-commit: 0bb9ec0aba2ecf536afd4c8cc03e103faceadca0