Add API fallback to UserDataContext (#8908)
Add a `returnDataFrom` property to `UserDataContext` that allows the
caller to decide how data from the context and data from the API
should be reconciled.  They can either solely use data from the
context or the API or say that both should be checked and one of them
preferred when they both return data.

Test Plan:

Added UserDataContext with an irrelevant user to docs thread.  With
`onlyContext`, it rendered like this:

<img width="341" alt="Screenshot 2024-04-24 at 4 06 59 PM" src="https://github.com/getcord/monorepo/assets/215991/ddab72eb-7b54-4351-b908-ae9c50f5adeb">

With any other value, it rendered like this:

<img width="327" alt="Screenshot 2024-04-24 at 4 07 47 PM" src="https://github.com/getcord/monorepo/assets/215991/e4f2476c-465a-49bd-b382-5e805fa6664b">

monorepo-commit: d41e1a2657d6794319a4e47fb97eff1b81eb00b6