Fix accidentally overriding className in `MessageFilesAttachments`
I was looking into why a Clack message had a huge `<video>`, when
ostensibly we have a `maxHeight` and `maxWidth` in
`sdk-js/packages/react/experimental/components/message/MessageVideoAttachment.css.ts`.
The issue is that, via `useCord`, this component was receiving a
`className` in `restProps` (despite its type not declaring that it
accepts such), which was overriding the `className` in the 3 attachment
wrapper `div`s` due to the `{...restProps}` spread, which would cause
the CSS selectors keeping the height/width in check not to match.

This component should properly accept a `className` -- and a `style`
while we're at it -- and apply them properly to each of the children
that it renders, without accidentally overriding their other classes.
This will allow the `maxHeight` and `maxWidth` we are already specifying
to actually match and do their job.


Reviewers: lazybean, flooey, AlbertQM

Reviewed By: AlbertQM, flooey

Pull Request: https://github.com/getcord/monorepo/pull/8926

monorepo-commit: 8832920f47de2503b728f4d241f8eafd242e7c46