First attempt at a streaming API endpoint.
This adds the incremental streaming update API. The idea
is we have an existing thread/message with a single markdown
node for content.  The call to this endpoint will append the
content supplied to the API to the end of the text of the
specified message. There are some guide rails to make
sure the message is in the expected format. Some unit
tests included to validate a happy path, and a couple
of the invalid formats.

Test Plan:
I've done a fairly adhoc set of manual tests, as well
as the included unit tests,  and trying now
to deploy to a windows machine for some sanity checking.


Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/8855

monorepo-commit: 465228d02b7633cdc55be9929b85f2ed355ec0d9