Message And Composer Layouts receives classnames (#8917)
Summary:
If the layout set the classnames, the the replacing layout will have to
set it as well.
It can be annoying.

For instance, we lose cord-message and a bunch of style break.
User currently has to find what is the missing class.

This changes make the parent pass the classname, so if you apply it you
are good.

Drawback, is you need a bit of massage if you actually want to remove
the classnames (see doc demos), but keep some others

Test Plan:
Docs thread/message/composer demos

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/72dccdb8-cdb8-4549-a036-f6e4c4b86841.png)

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/IoyKnJXsDnJREClofLDY/26e66cef-b1b1-41a4-a75c-f1b8243e7f10.png)


Reviewers:

monorepo-commit: 1eb8cc681d1baa56d3835a4882bfc116720d34f3