TypingIndicator:Typing in SendComposer set typing to true (#8910)
Summary:
When user types in SendComposer we update the thread to set `{typing:
true}`

Note: this does not implement the UI indicator at all. Only signalling
cord that we type.

I do not do that in EditComposer because it could be confusing.
The typing indicator expires by itself, there is no need to do it.
(In fact I initially did it onBlur and it did not seem to do it more
quickly that letting it expire)

Test Plan:
2 test plan side by side.
With nonion Thread, type.
With onion Thread see the typing indicator.

Reviewers:

monorepo-commit: 740df2172d6375a83c56c4068c95cc2ea55af9dc