Composer 4.0: add `expanded` prop
This property allows developers to control when the composer expands.

The prop is reflected 1:1 in the class names.
The inline composer is created with just CSS: this removes some annoyances with Slate (e.g. `useFocused()` and `ReactEditor.isFocused` don't work). It's also nicer to _not_ add more state to the already complex composer.

Test Plan:
Check that focusing a composer will expand it.
Check that setting prop to `never` will never expand it, and setting it to `always` will never collapse it.


Reviewers: lazybean, elenaperezrioja, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8949

monorepo-commit: dfaa76715c1ae4e2fa170f791c26012740b5d97d