SendComposer: Add `showCancelButton` prop
The `SendComposer` can collapse (#8949). In some cases, we want the collapsed composer to have a X button, so we can do some action.
An example of this is in Threads: you click `Reply`, we now show a collapsed composer. If you click X, we get rid of it. Without the X button, there would be no way of hiding it.


Test Plan:
Add `showCancelButton` to the `SendComposer` in the testbed, see the X button.

Reviewers: Netceer, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8957

monorepo-commit: a83e810a71620ebe99f8512eaa5509e4aeee449c