Thread 4.0 components: Fix margin
Each individual component of a thread did not have any CSS, but it should.
This resulted in the comopser, scrollcontaner, etc to be as wide as the thread container, which looked funky -- so funky we added a `12px` padding everywhere. I'm undoing this padding, and adding the proper margin/padding to each component

Test Plan:
Load docs locally. Thread looks good now
<img width="334" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/c0cef20c-71bd-4df8-8aa4-94b7edb694e9">
<img width="364" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/7978414a-bbb1-4df5-9af2-5f2d34a572ee">
<img width="357" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/3482f3fa-952f-46aa-aba9-21272ec417d6">



Reviewers: Netceer, mhtien, lazybean, elenaperezrioja

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8981

monorepo-commit: 708f44a5de0a803e248f673654f1c05e11bef53f