Add and expose ComposerAttachmentInput component in betaV2
I had an issue with trying to put the AddAttachment
button into a menu (for the docs CompactComposer example)
and I couldn't get it working because I needed to
close the menu, and when you do this, the AddAttachment
component gets unmounted and then the attachment is gone.

This PR does the following:
1. Extracts the input element to be a replacement element
(so users can replace it and restrict file extensions etc)
2. Relocates the element from the AddAttachments button
so its a sibling to the ComposerLayout.


Reviewers: elenaperezrioja, Netceer, AlbertQM, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8971

monorepo-commit: 031ce8a9cbdc4694dadb3c68b8ac8d0cddd9f8dc