Remove composerShowCancelButton prop
Instead of having to expose `composerShowCancelButton` in every component that uses a composer, we are option to NOT expose the prop.
Instead, developers can replace the `Composer` `within` that `Thread` or `Message` and add the `showCancelButton` prop to the `Composer` themselved.

Test Plan:
Try replacing a Composer and passing the `showCancelButton` prop. It works.


Reviewers: elenaperezrioja, lazybean

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/8985

monorepo-commit: de1bee654d435c368b04eb19ad50dd4f43a4a498