Consolidate hooks that call observe functions (#8986)
There was a problem in the hooks where changing the input value to a
hook (such as an ID or the set of options) would continue returning
data from any previous callback while it waited for the new one to
come in.  Since fixing that in place would require changing like a
dozen implementations, I figured it was time to consolidate them,
since they all have the same structure.

Add a single `useObserveFunction` implementation that does the same
thing as all the custom functions, but in a generic way.  The generics
on the types are a little gnarly, but they should prevent a caller
from providing data that won't work when it goes to actually call the
observe function or process the return value.

Test Plan:

Visited every newstyle component in the docs, they all loaded properly
and received live updates.

Added a ThreadedComments with a value for `highlightThreadID` and a
button that changed to pass in the empty string instead.  Clicking the
button unhighlighted the thread (while before, it left it
highlighted).

Updated	Clack with these packages and used the features,
including viewing notifications, marking them as read, viewing
different channels, and sending a DM.  Everything updated
properly.

monorepo-commit: f821033a00c05f60e3d6a63d060d01baa6a5267a