Expose addAttachmentFilesToComposer in Composer beta component
There currently isn't a way for someone to attach
a file to the composer without using the add attachment
button that alreayd exists.

Exposing this function will allow someone to now
attach files they want to be shown within the composer.

I've just exposed an existing function we alreayd
had, but changes the arg to File[] rather than
FileList as this is a read-only type within
the input tag.

An exmaple of this (which will live in a separate PR)
is having a gif integration.

Wasn't sure what the best way is to add this function
in but there will be some example code of how to use it
(via the composer replacement, or composer layout).

Test Plan:
Will test in a later PR (for gif integration).


Reviewers: lazybean, AlbertQM, elenaperezrioja, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8995

monorepo-commit: 297021ef4fc29dfe9f97b51ea651f40e88bf3258