Add subscribeToThread to creating messages (#8999)
Add a parameter that defaults to true that subscribes the sender of a
message to the thread.  This is true by default, but can be disabled
in case this isn't desired.

Mostly this diff is just plumbing.


Test Plan:

npm run test

Send a message with the beta Cord composer, it properly subscribes the
user to the thread.

monorepo-commit: cf36d959c83ca0eda3cac47a449ec611f6e76cb2