New `Threads` component
We know we need a component that gets a list of `threads` and renders them, as most of our clients use one.

The rough plan is to start with a simple one (like the one in this PR), and use this as a building block to create more complex UIs (e.g. tabbed view of resolved / unresolved).

This components needs more work: e.g. exposing Threads.ByLocation (or similar). These enhancements will come in future PRs.

Test Plan:
Play with the component in the testbed.


Reviewers: flooey, elenaperezrioja, Netceer, mhtien

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/8984

monorepo-commit: a85ebbe9e10f6d1dd1ec0b5ff7c8fe6f4c59652a