Composer 4.0: Allow sending attachments not yet uploaded
The `updateAttachment` function adds a `type: 'file'` to the file attached by the user. This turns said file into an actual attachment that will be sent as part of the message.

The first invocation of `updateAttachment` happens when we tell the server to start uploading the file.
However, this initial call was NOT adding the `type: 'file'`. This was added later on, when the upload finished and we heard back from the server.
So, if you sent a message whilst the file you attached was still uploading, it would send an empty message.

This PR makes sure we add the `type: 'file'` on the first invokation of `updateAttachment` (which we fire as an optimistic update, before the file is actually uploaded onto S3).

Test Plan:
Upload a file.
Send message before it has finished uploading.
It correctly sends a message with the attachment (which will finish uploading and will be shown on the page -- not as a preview though. You have to refresh to see the preview).

Upload a file.
Cancel upload.
Type hello and send. Can send message, and it has no attachment.

Upload a file.
Let it finish uploading.
Send message. All works.


Reviewers: elenaperezrioja, lazybean, jwatzman, Netceer

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/9009

monorepo-commit: dab9f695b24799439871df646c6b4ee7a7cb4c24