Upgrade rollup and friends to current version
There's some fixing with our packaging I'm looking into, and I want to
play with some features available with the newest rollup. So let's
upgrade.

This is two major versions forward. Changelogs for the breaking changes
are here:
https://github.com/rollup/rollup/blob/master/CHANGELOG.md#300
https://github.com/rollup/rollup/blob/master/CHANGELOG.md#400

Honestly, I didn't read through them all that carefully. What I did do
was generate `dist` for our `chatbot-base` (because it is simple) and
`react` (because it is not) packages, and `diff` them before and after
this change. In the mjs files, the only change was around comments, and
a couple of objects have fields in slightly different order, nothing
interesting. For cjs, the changes were mostly some renaming etc,
although we also will no longer be shipping the `__esModule: true` sigil
-- which doesn't matter since we aren't doing default exports. (This is
due to one of the changes in 3.0.0, changing the default for
`output.esModule`, see changelog above.) Basically none of the changes
seemed to matter.

I also injected the new version of these packages into Clack. It uses
mjs by default, which ran fine. Forcing it to use the cjs version of the
React package caused no more issues than forcing it to use the cjs
version of that package before this change (the composer fails to render
for both, likely due to some ordering change tickling a Clack bug, but
it's not a *new* break).

So I think we're good.

Ref getcord/sdk-js#1


Test Plan: As above.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/9028

monorepo-commit: a866c6b1640a81d19bae68a0d9664c11f8fd203e