Introduce rollup-plugin-cord-pure
We have a serious problem with tree-shake-ability right now, especially
with the new Cord 4.0 components. Many/all of them are defined like
this:

```
export const CoolThing = withCord(forwardRef( ... ));
```

Even if an end-user of the Cord SDK never uses `CoolThing` in their
actual code, their bundler can't tree-shake it because it can't know
that `withCord` and `forwardRef` don't have side effects or otherwise
mutate global state. So it has to keep `CoolThing` -- and everything it
transitively depends on! -- in the final bundle.

This is a real problem -- just adding a `CordProvider` -- literally
nothing else from the SDK, just that -- adds 1.7M to our customers'
uncompressed bundle size (as of version 1.43.1)! Customers have noticed
and complained, cf. getcord/sdk-js#1.

But these functions *don't* actually have side-effects. Thankfully there
is a way to annotate that -- two, in fact:

1. `@__PURE__`, which goes at a callsite, which says "this particular
   call is safe to tree-shake away if needed"
2. `@__NO_SIDE_EFFECTS`, which goes on a function definition, which says
   "all calls to this function are always pure and safe to tree-shake
   away if needed"

Option (2) is really nice for things like `withCord`, so that we don't
have to go and annotate every single callsite (and inevitably miss one).
Unfortunately esbuild doesn't really support it -- it will keep the
annotation, but won't do anything with it, because esbuild doesn't/can't
do the sort of global analysis that's needed to make it work. And so
while (2) is nice for us, it would be unfortunate if our customers using
esbuild weren't able to tree-shake because we used an incompatible
annotation!

But, esbuild does support (1) (since it is completely local), and we
ourselves *do* global analysis via Rollup when we build our bundle to
ship to npm. So we can use a Rollup plugin to transform (2) into (1) at
build time.

Unfortunately, AFAICT no such plugin exists. The closest is this:
https://github.com/danielroe/rollup-plugin-pure but it's regex based,
and worse it doesn't actually transform (2) into (1), rather it uses a
list passed via build configuration of where to add the annotations for
(1). We could make this work, but having such an important annotation
effectively live far away from the function definition, in a place most
people don't look, which would break in a subtle-but-important way if,
e.g., `withCord` were to get renamed -- that seemed bad.

And it turns out that it's not *that* hard to write a simple version of
a plugin that transforms (2) into (1) since it doesn't need to be
general or super robust as long as it works for the things the Cord SDK
actually does in practise. So I did. (It also accepts a hardcoded list
of other identifiers to mark as always pure, for library functions like
`React.forwardRef`.)


Test Plan:
The above plugin, with the one annotation and 3 hardcoded identifiers
included here, saves 645k uncompressed bundle size from my above test
case that uses only `CordProvider`. Still a long way to go, but it's a
fantastic start. Inspecting the generated cjs and mjs for the React
package looks like annotations are in the right place. Injecting the
SDKs into Clack doesn't break it.

Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM, flooey

Pull Request: https://github.com/getcord/monorepo/pull/9037

monorepo-commit: 0cf992234cce2c2c917eb6059756674d3637b84b