Remove noReactions css from beta Messsage
I noticed in the test bed the options menu were not aligning properly
when there were no reactions.
![image](https://github.com/getcord/monorepo/assets/59407030/c485825e-4a8b-43f2-b8be-c80cd748aa02)

When there was a reaction it would align properly (i.e. to the right).
![image](https://github.com/getcord/monorepo/assets/59407030/64e0793a-b8ff-491e-a841-11a41de47bfb)

I'm not sure if this happened because we still have the reactions container
within the grid, and it's got a grid name...?

Alternatively is we keep this CSS, but do not render the reactions container
if not reactions exist on the message.

Test Plan:
Options menu is aligned now.
<img width="559" alt="image" src="https://github.com/getcord/monorepo/assets/59407030/ca63fd8b-dc6a-4669-abd1-13b9b0f255ad">



Reviewers: lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9058

monorepo-commit: 7fd2519876b1c254b583cd227c20145d4bfc9e48