Add Drag and drop to new composer
Predominantly copied over code from ui3 - and
decided to keep it as the composer container (similar
to ui3 as opposed to having a separate div container
within composer layout but this can always be discussed)

Becuase of this, I thought it made sense to allow
users to disable the drag and drop so I added it
to the Composer as a prop.

There are two new things - WithDragAndDrop, and DragAndDrop.

WithDragAndDrop - should be used if the user wants
to replace the ComposerLayout and keep the drag and
drop functionality.

DragAndDrop is a replacable, if users wanted to log
extra things in the drag events.

Will update the doc examples to show use of WithDragAndDrop


Reviewers: AlbertQM, elenaperezrioja

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/9025

monorepo-commit: 97adb20be5b114dcfa29ac45ce3df952b488d75b