Have explicit skip value for React hooks (#9086)
We were returning the loading value for React hooks when the skip
option was supplied.  For hooks where the loading value is undefined,
this makes sense, but some hooks return a structured object with
`loading: true`, and that didn't make sense if you're skipping (it's
definitely not loading).  Add the ability to supply a skip value in
that case, and supply a `loading: false` version for all those hooks.

Test Plan:

Added a useThreads with a skip and a button to swap the skip value,
and then logged all the return values.  When skip was true, always
received the skip value, otherwise got the loading or data values as
appropriate.

monorepo-commit: 92e4a03a0a2d67e0fe6a8abaf1f2a72be884d17d