Delay loading editor.children until ready to send
There's a subtle bug in #8996 -- it shifts the load of `editor.children`
into the React render cycle, instead of when the callback is actually
called. I guess we don't rerender when `editor.children` changes (quite
reasonably), so since that is outside of what React knows about, we need
to be careful when exactly we load it. Refactor and shift it back into
the callbacks.

Note that we can't factor out this helper const to toplevel for the same
reason:

```
const message = {
  ...initialValue,
  ...value,
  content: editor.children,
};
```


Test Plan: Bug in thread mentioned above no longer repros.

Reviewers: AlbertQM, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9095

monorepo-commit: 6ee89f81cbdde91f1f894cad4e171a48f898e245