ThreadProps: Make threadData required
Not sure why this was marked as optional (with a `?`) at the top level, and required but possibly `undefined` down below (`: ClientThreadData | undefined`).

ThreadData is never undefined. What the API returns is an object with a bunch of fields, some of which can be undefined while loading the data.

Simplifying the code here

Test Plan:
Load testbed. Threads still load as expected.

Reviewers: lazybean, Netceer

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/9072

monorepo-commit: 8f92596c874c883c91e3a532bbaa43d25fa95c26