Threads Layouts: Collapsed and Expanded
Before this PR, we were tracking two states: expanded and showComposer.
The former controlled whether we were showing replies or not; the latter whether we were showing the composer or not.
This was a bit complex, and all this logic lived in the InlineThreadLayout. This made it very hard to replace this component, as you'd have to rewrite all that logic yourself.

Instead, we're simplifying the Layout to either be "Expanded" or "Collapsed" --  making the UX of this component simpler, which in turn will make devs life easier when trying to customise it.

"Collapsed": we only show the first message, and an expand button to "Reply" or "X replies..".
"Expanded": we show the first message, the collapse button, the other messages and the composer.

Note: you now cannot "hide" the composer if the InlineThread isExpanded. You can however collapse it and the composer will go away.

Test Plan:
Load Threads.
Threads without reply say "Reply". Clicking on it will go into "Expanded", showing a composer and the "Hide replies" button. (might want to change this label).

Reply to thread, now it has 1 reply. Clicking "Hide replies" will go back to "Collapsed". Button now says "1 Reply", and there's no composer. Expanding it again shows the composer as expected.


Reviewers: lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9080

monorepo-commit: 70de2ac42c8968a7e37c9556067480c4bb512d43