Passing props to the composer in Threads (#9105)
## Summary

When using Threads in clack, it is not possible to control where the
messages created in the composer end up. We are only passing the group
ID to the composer, but not the location of the thread created or any
other data (metadata, subscribers,...). I think we should pass this info
in the composerOptions, so that they're passed down to the composer.

monorepo-commit: 5261bad3db922efb3b581e635d4dfba56a74ec38