ScrollContainer: Use MutationObserver rather than children.length
The original ScrollContainer was super simple: it relied on `children.length` inside the dependency array to know where to re-run some logic. However, in some cases, this is not enough -- for the typing indicator, we _always_ pass it as a `child`, but it is sometimes hidden (removed) in its render function. This meant the `children.length` didn't change, though the indicator was actually removed from the DOM.
To detect this, the only way is to use a `MutationObserver`.

This addition simplifies replacing `ThreadLayout` too, as we now don't need to pass an array of `children`, but just the components themselves.

Test Plan:
Add lots of `console.log` checking when is the observer created, when does it start `observe`, and when is the callback function called.

Load testbed, Thread correctly scrolls to the bottom if I'm at the bottom edge. Keeps scrolling the same if I've scrolled up.


Reviewers: Netceer, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/9094

monorepo-commit: 217775db6df24959089728da2ccf88a8f44212f6